[MAINT] Stop unnecessary dtype conversions (#154)

+++ src/pybispectra/utils/utils.py
@@ -67,11 +67,7 @@ def compute_fft(
     if verbose:
         print("Computing FFT on the data...")
 
-    freqs = (
-        sp.fft.rfftfreq(n=n_points, d=1 / sampling_freq)
-        .astype(np.float32)  # reduce rounding errors
-        .astype(_precision.real)
-    )
+    freqs = sp.fft.rfftfreq(n=n_points, d=1 / sampling_freq).astype(_precision.real)
 
     window = window_func(data.shape[2])
 
@@ -453,7 +449,7 @@ _DATA_ALIAS_FILE_HASH = [
 
 DATASETS = {alias: filename for alias, filename, _ in _DATA_ALIAS_FILE_HASH}
 
-if "dev" not in version:
+if "dev" not in version:  # pragma: no cover
     version = Version(version)
     version = f"{version.major}.{version.minor}"  # take data from maj.min branch
